pyproject.toml
[project]
name = "swjtu-score-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "ddddocr==1.0.6",
    "fastapi>=0.124.4",
    "pillow==9.5.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
]
